+++ src/matoid/matoid.ts
@@ -1,0 +1,7 @@
+export {assert} from './assert.js'
+export {Int} from './int.js'
+export type {Millis} from './millis.js'
+export {NumberUtil} from './number-util.js'
+export {Rect, RInt} from './rect.js'
+export {WH, WHInt} from './wh.js'
+export {XY, XYInt} from './xy.js'