+++ docs/changelog.md
@@ -4,4 +4,5 @@ Versions and bullets are arranged chronologically from latest to oldest.
 
 ## v0.0.1 (unreleased)
 
+- \[build\]\[API\] Rename export from "math" to "matoid"
 - Initial release